requirements.txt
discord.py==2.3.2
groq>=0.9.0
